+++ lib/radar/reporter/hoptoad_reporter.rb
@@ -182,7 +182,7 @@ module Radar
             {}
           @_request_info[:controller] = @_request_info[:parameters]['controller']
           @_request_info[:action]     = @_request_info[:parameters]['action']
-          @_request_info[:cgi_data]   = hash[:request][:rack_env] || {}
+          @_request_info[:cgi_data]   = hash[:request][:rack_env] || hash[:request][:headers] || {}
           @_request_info[:session]    = hash[:request][:rack_env]['rack.session'] || {}
         end